Use generate_far_filename to generate cache filenames

+++ nemo_text_processing/inverse_text_normalization/hy/verbalizers/verbalize_final.py
@@ -13,12 +13,21 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import os
+
 import pynini
 from pynini.lib import pynutil
 
 from nemo_text_processing.inverse_text_normalization.hy.verbalizers.verbalize import VerbalizeFst
 from nemo_text_processing.inverse_text_normalization.hy.verbalizers.word import WordFst
-from nemo_text_processing.text_normalization.en.graph_utils import GraphFst, delete_extra_space, delete_space
+from nemo_text_processing.text_normalization.en.graph_utils import (
+    GraphFst,
+    delete_extra_space,
+    delete_space,
+    generator_main,
+    generate_far_filename,
+)
+from nemo_text_processing.utils.logging import logger
 
 
 class VerbalizeFinalFst(GraphFst):
@@ -27,19 +36,46 @@ class VerbalizeFinalFst(GraphFst):
     tokens { name: "նա" } tokens { cardinal { integer: "23" } } tokens { name: "տարեկան" } tokens { name: "է" } -> նա 23 տարեկան է
     """
 
-    def __init__(self, project_input: bool = False):
-        super().__init__(name="verbalize_final", kind="verbalize")
-        verbalize = VerbalizeFst(project_input=project_input).fst
-        word = WordFst(project_input=project_input).fst
-        types = verbalize | word
-        graph = (
-            pynutil.delete("tokens")
-            + delete_space
-            + pynutil.delete("{")
-            + delete_space
-            + types
-            + delete_space
-            + pynutil.delete("}")
-        )
-        graph = delete_space + pynini.closure(graph + delete_extra_space) + graph + delete_space
-        self.fst = graph
+    def __init__(
+        self,
+        deterministic: bool = True,
+        project_input: bool = False,
+        cache_dir: str = None,
+        overwrite_cache: bool = False
+    ):
+        super().__init__(name="verbalize_final", kind="verbalize", deterministic=deterministic, project_input=project_input)
+
+        far_file = None
+        if cache_dir is not None and cache_dir != "None":
+            os.makedirs(cache_dir, exist_ok=True)
+            far_file = generate_far_filename(
+                language="hy",
+                mode="itn",
+                cache_dir=cache_dir,
+                operation="verbalize",
+                deterministic=deterministic,
+                project_input=project_input
+            )
+        
+        if not overwrite_cache and far_file and os.path.exists(far_file):
+            self.fst = pynini.Far(far_file, mode="r")["verbalize"]
+            logger.info(f'VerbalizeFinalFst graph was restored from {far_file}.')
+        else:
+            verbalize = VerbalizeFst(project_input=project_input).fst
+            word = WordFst(project_input=project_input).fst
+            types = verbalize | word
+            graph = (
+                pynutil.delete("tokens")
+                + delete_space
+                + pynutil.delete("{")
+                + delete_space
+                + types
+                + delete_space
+                + pynutil.delete("}")
+            )
+            graph = delete_space + pynini.closure(graph + delete_extra_space) + graph + delete_space
+            
+            self.fst = graph.optimize()
+            if far_file:
+                generator_main(far_file, {"verbalize": self.fst})
+                logger.info(f"VerbalizeFinalFst grammars are saved to {far_file}.")
